submissions/devikahooda775/level6/requirements.txt
streamlit
neo4j
python-dotenv
pandas
plotly